Move snapshot related classes to another namespace
