+++ Source/EventFlow.TestHelpers/Aggregates/ThingyAggregate.cs
@@ -27,8 +27,8 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using EventFlow.Aggregates;
-using EventFlow.EventStores.Snapshots;
 using EventFlow.Exceptions;
+using EventFlow.Snapshots;
 using EventFlow.TestHelpers.Aggregates.Entities;
 using EventFlow.TestHelpers.Aggregates.Events;
 using EventFlow.TestHelpers.Aggregates.Snapshots;
