+++ Source/EventFlow.Tests/UnitTests/EventStores/Snapshots/SnapshotUpgradeServiceTests.cs
@@ -26,8 +26,8 @@ using System.Threading;
 using System.Threading.Tasks;
 using EventFlow.Configuration;
 using EventFlow.Core.VersionedTypes;
-using EventFlow.EventStores.Snapshots;
 using EventFlow.Logs;
+using EventFlow.Snapshots;
 using EventFlow.TestHelpers;
 using EventFlow.TestHelpers.Aggregates.Snapshots;
 using EventFlow.TestHelpers.Aggregates.Snapshots.Upgraders;
