+++ Source/EventFlow/EventFlow.csproj
@@ -147,26 +147,26 @@
     <Compile Include="EventStores\IEventPersistence.cs" />
     <Compile Include="EventStores\IEventUpgradeManager.cs" />
     <Compile Include="EventStores\IEventUpgrader.cs" />
-    <Compile Include="EventStores\InMemory\InMemorySnapshotPersistence.cs" />
-    <Compile Include="EventStores\Snapshots\CommittedSnapshot.cs" />
-    <Compile Include="EventStores\Snapshots\ISnapshotBuilder.cs" />
-    <Compile Include="EventStores\Snapshots\ISnapshotMetadata.cs" />
-    <Compile Include="EventStores\Snapshots\ISnapshotPersistence.cs" />
-    <Compile Include="EventStores\Snapshots\ISnapshotUpgradeService.cs" />
-    <Compile Include="EventStores\Snapshots\SerializedSnapshot.cs" />
-    <Compile Include="EventStores\Snapshots\SnapshotContainer.cs" />
-    <Compile Include="EventStores\Snapshots\SnapshotBuilder.cs" />
-    <Compile Include="EventStores\Snapshots\SnapshotMetadata.cs" />
-    <Compile Include="EventStores\Snapshots\SnapshotMetadataKeys.cs" />
-    <Compile Include="EventStores\Snapshots\SnapshotStore.cs" />
+    <Compile Include="Snapshots\Stores\InMemory\InMemorySnapshotPersistence.cs" />
+    <Compile Include="Snapshots\CommittedSnapshot.cs" />
+    <Compile Include="Snapshots\ISnapshotBuilder.cs" />
+    <Compile Include="Snapshots\ISnapshotMetadata.cs" />
+    <Compile Include="Snapshots\ISnapshotPersistence.cs" />
+    <Compile Include="Snapshots\ISnapshotUpgradeService.cs" />
+    <Compile Include="Snapshots\SerializedSnapshot.cs" />
+    <Compile Include="Snapshots\SnapshotContainer.cs" />
+    <Compile Include="Snapshots\SnapshotBuilder.cs" />
+    <Compile Include="Snapshots\SnapshotMetadata.cs" />
+    <Compile Include="Snapshots\SnapshotMetadataKeys.cs" />
+    <Compile Include="Snapshots\Stores\SnapshotStore.cs" />
     <Compile Include="EventStores\ISerializedEvent.cs" />
-    <Compile Include="EventStores\Snapshots\ISnapshot.cs" />
-    <Compile Include="EventStores\Snapshots\ISnapshotDefinitionService.cs" />
-    <Compile Include="EventStores\Snapshots\ISnapshotStore.cs" />
-    <Compile Include="EventStores\Snapshots\SnapshotDefinition.cs" />
-    <Compile Include="EventStores\Snapshots\SnapshotUpgradeService.cs" />
-    <Compile Include="EventStores\Snapshots\SnapshotVersionAttribute.cs" />
-    <Compile Include="EventStores\Snapshots\SnapshotDefinitionService.cs" />
+    <Compile Include="Snapshots\ISnapshot.cs" />
+    <Compile Include="Snapshots\ISnapshotDefinitionService.cs" />
+    <Compile Include="Snapshots\Stores\ISnapshotStore.cs" />
+    <Compile Include="Snapshots\SnapshotDefinition.cs" />
+    <Compile Include="Snapshots\SnapshotUpgradeService.cs" />
+    <Compile Include="Snapshots\SnapshotVersionAttribute.cs" />
+    <Compile Include="Snapshots\SnapshotDefinitionService.cs" />
     <Compile Include="Exceptions\DomainError.cs" />
     <Compile Include="Exceptions\DuplicateOperationException.cs" />
     <Compile Include="Exceptions\MetadataKeyNotFoundException.cs" />
