+++ Source/EventFlow/Snapshots/ISnapshot.cs
@@ -1,7 +1,7 @@
 ﻿// The MIT License (MIT)
 // 
-// Copyright (c) 2015-2016 Rasmus Mikkelsen
-// Copyright (c) 2015-2016 eBay Software Foundation
+// Copyright (c) 2015 Rasmus Mikkelsen
+// Copyright (c) 2015 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -22,11 +22,11 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-namespace EventFlow.EventStores.Snapshots
+using EventFlow.Core.VersionedTypes;
+
+namespace EventFlow.Snapshots
 {
-    public sealed class SnapshotMetadataKeys
+    public interface ISnapshot : IVersionedType
     {
-        public const string SnapshotName = "snapshot_name";
-        public const string SnapshotVersion = "snapshot_version";
     }
 }
