+++ Source/EventFlow/Snapshots/Stores/InMemory/InMemorySnapshotPersistence.cs
@@ -22,23 +22,9 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // 
 
-using System.Threading.Tasks;
-using EventFlow.Aggregates;
-using EventFlow.Core;
-
-namespace EventFlow.EventStores.Snapshots
+namespace EventFlow.Snapshots.Stores.InMemory
 {
-    public interface ISnapshotStore
+    public class InMemorySnapshotPersistence : ISnapshotPersistence
     {
-        Task<TSnapshot> LoadSnapshotAsync<TAggregate, TIdentity, TSnapshot>(
-            TIdentity identity)
-            where TAggregate : IAggregateRoot<TIdentity>
-            where TIdentity : IIdentity
-            where TSnapshot : ISnapshot;
-
-        Task StoreSnapshotAsync<TAggregate, TIdentity, TSnapshot>()
-            where TAggregate : IAggregateRoot<TIdentity>
-            where TIdentity : IIdentity
-            where TSnapshot : ISnapshot;
     }
 }
